fix(gql): Missing billing_configuration solution for OrganizationType
